rename GameNameDuplicationsSpecification
